backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
geopy
email-validator
